+++ multiplatform-crypto-libsodium-bindings/src/jvmMain/kotlin/com/ionspin/kotlin/crypto/stream/Stream.kt
@@ -32,7 +32,7 @@ actual object Stream {
     actual fun chacha20IetfXorIc(
         message: UByteArray,
         nonce: UByteArray,
-        initialCounter: ULong,
+        initialCounter: UInt,
         key: UByteArray
     ): UByteArray {
         val result = UByteArray(message.size)
@@ -42,7 +42,7 @@ actual object Stream {
             message.asByteArray(),
             message.size.toLong(),
             nonce.asByteArray(),
-            initialCounter.toLong(),
+            initialCounter.toInt(),
             key.asByteArray()
         )
